requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
qrcode==7.3.1
six==1.16.0
